requirements.txt
pinecone==6.0.2      
openai==1.52.0          
python-dotenv==0.21.0   
requests==2.32.3      
